feat(cli): Throw if Lowdefy CLI is run with node < v14.
