+++ packages/cli/src/index.js
@@ -14,106 +14,16 @@
   See the License for the specific language governing permissions and
   limitations under the License.
 */
+const nodeMajorVersion = process.version.split(/^v(\d+)/)[1];
+if (Number(nodeMajorVersion) < 14) {
+  // TODO: This error handled with telemetry.
+  throw new Error(
+    `Nodejs versions below v14 are not supported. You are using ${process.version}. Update Nodejs to the latest LTS version to use Lowdefy.`
+  );
+}
+async function run() {
+  const { default: program } = await import('./program.js');
+  program.parse(process.argv);
+}
 
-import { createRequire } from 'module';
-import { Command } from 'commander';
-
-import build from './commands/build/build.js';
-import dev from './commands/dev/dev.js';
-import init from './commands/init/init.js';
-import start from './commands/start/start.js';
-import runCommand from './utils/runCommand.js';
-
-const require = createRequire(import.meta.url);
-
-const packageJson = require('../package.json');
-const { description, version: cliVersion } = packageJson;
-
-const program = new Command();
-
-program.name('lowdefy').description(description).version(cliVersion, '-v, --version');
-
-program
-  .command('build')
-  .description('Build a Lowdefy production app.')
-  .usage(`[options]`)
-  .option(
-    '--config-directory <config-directory>',
-    'Change config directory. Default is the current working directory.'
-  )
-  .option('--disable-telemetry', 'Disable telemetry.')
-  .option('--no-next-build', 'Do not build the Next.js server.')
-  .option(
-    '--package-manager <package-manager>',
-    'The package manager to use. Options are "npm" or "yarn".'
-  )
-  .option(
-    '--ref-resolver <ref-resolver-function-path>',
-    'Path to a JavaScript file containing a _ref resolver function to be used as the app default _ref resolver.'
-  )
-  .option(
-    '--server-directory <server-directory>',
-    'Change the server directory. Default is "<config-directory>/.lowdefy/server".'
-  )
-  .action(runCommand({ cliVersion, handler: build }));
-
-program
-  .command('dev')
-  .description('Start a Lowdefy development server.')
-  .usage(`[options]`)
-  .option(
-    '--config-directory <config-directory>',
-    'Change config directory. Default is the current working directory.'
-  )
-  .option('--disable-telemetry', 'Disable telemetry.')
-  .option('--no-open', 'Do not open a new tab in the default browser.')
-  .option(
-    '--package-manager <package-manager>',
-    'The package manager to use. Options are "npm" or "yarn".'
-  )
-  .option('--port <port>', 'Change the port the development server is hosted at. Default is 3000.')
-  .option(
-    '--ref-resolver <ref-resolver-function-path>',
-    'Path to a JavaScript file containing a _ref resolver function to be used as the app default _ref resolver.'
-  )
-  .option(
-    '--watch <paths...>',
-    'A list of paths to files or directories that should be watched for changes. Globs are supported. Specify each path to watch separated by spaces.'
-  )
-  .option(
-    '--watch-ignore <paths...>',
-    'A list of paths to files or directories that should be ignored by the file watcher. Globs are supported. Specify each path to watch separated by spaces.'
-  )
-  .option(
-    '--dev-directory <dev-directory>',
-    'Change the development server directory. Default is "<config-directory>/.lowdefy/dev".'
-  )
-  .action(runCommand({ cliVersion, handler: dev }));
-
-program
-  .command('init')
-  .description('Initialize a Lowdefy project.')
-  .usage(`[options]`)
-  .action(runCommand({ cliVersion, handler: init }));
-
-program
-  .command('start')
-  .description('Start a Lowdefy production app.')
-  .usage(`[options]`)
-  .option(
-    '--config-directory <config-directory>',
-    'Change config directory. Default is the current working directory.'
-  )
-  .option('--disable-telemetry', 'Disable telemetry.')
-  .option(
-    '--package-manager <package-manager>',
-    'The package manager to use. Options are "npm" or "yarn".'
-  )
-  .option('--port <port>', 'Change the port the server is hosted at. Default is 3000.')
-  .option(
-    '--server-directory <server-directory>',
-    'Change the server directory. Default is "<config-directory>/.lowdefy/server".'
-  )
-  .action(runCommand({ cliVersion, handler: start }));
-
-program.parse(process.argv);
+run().then(() => {});
